Move code signing and path mangling to Binary member functions run at copy time.
To ensure that binaries are signed bottom-up and to provide for future
path specialization to meet Apple bundle format requirements.
A rather wide-ranging change, because it requires accumulating the dependencies
before beginning the copy (because the copied binaries have mangled dependency
rpaths which are more difficult to use). Neither codesign nor otool take
kindly to being handed directories, so directory project entries need to be
handled a bit differently as well.

The copying of the main binary is moved to be the last thing (or next-to-last
if there’s a launcher script) so that all of the libraries and loadable
modules signatures are incorporated in the main binary’s signature.