Copy data sections before copying binaries
Copying binaries has secondary effects such as run-install-name-tool,
and if any of the specified binaries are overwritten by a data section,
those effects will be overwritten as well. An example from my project:

    <binary>${prefix}/lib/python2.7/lib-dynload/*.so</binary>
    <data>${prefix}/lib/python2.7</data>

In this case, we want to grab all the python modules, and make all the
dynamically loaded binary modules relocatable with
run-install-name-tool. But since data files are copied later than
binaries, the relocated binaries are overwritten.

There are other solutions: we could enumerate all the data files
individually or by glob instead of specifying the whole prefix. The
copy_target function could avoid overwriting existing files. But this
was the simplest option that worked for me.