bundler.py: Ensures dependent binaries are copied
Without this change, `dylib`s that other `dylibs` depend on are not
copied. This ensures that they are so that the full tree is copied.

It does appear, when tracing the copies, that quite a few `dylib`s are
copied more than once. A simple step of removing duplicates should
ensure this stops happening, but has not been done yet.